ai_email_rewriter/requirements.txt
txt
requests
tkinter
python-dotenv
